Search for the second largest number
Find out the unique second largest number from an array even if the array has repetitive number elements. I sorted out the array first then I reverse searched and compared the elements from the last to the unique second-largest number as long as it's not equal to the first largest number.